requirements.txt
colorama
anthropic
pygments
python-dotenv
Pillow
google-generativeai
